feat: 838 text encoder/decoder implemented
